Introduce first blog post
It's about the history of DevRel at Tailscale and how I used a bit of
alchemy to transmute a part of myself that I had come to hate into a
powerful force of good that helps me get my job done.

Also introduces a HeroImage component to get images into the mix for
breaking up large blocks of text.

Signed-off-by: Xe Iaso <xe@tailscale.com>