.github/workflows: [fly.yml] change deploy to `--local-only`
Changing to `flyctl deploy --local-only` to try and improve deploy
times. We should be able to use GitHub Actions to build the container
and just push it to Fly.io. This should be faster since we're not
waiting on their builder to be available instead, we're using the
existing runner to do the build.

Updates: #36