lib/utils: [contentlayer] sort events in ascending order
Events should be ascending order with the most recent event first. Blog
posts are opposite.

I removed the `dateSortDesc` function because it wasn't really date
specific. It was also exported but not used anywhere else.

I inlined the sort and parse the date so we can do `a-b`, `b-a` sorting.

Fixes: #29

Signed-off-by: Shayne Sweeney <shayne@tailscale.com>