Add go reverse proxy for the website
This will be used more in the future for pure-API things and
integrations with third parties like Slack. This includes bindings and
the like to Tailscale's version of Go so that everyone can have a
consistent toolchain.

This also does some Nix magic that I copied from the tailscale/tailscale
flake.nix to have Nix manage the right version of Go instead of having
to download and build Go separately.

The Dockerfile, all development commands, and more all do the correct
things and things should be otherwise unrecognizable.

The Dockerfile also builds a correct static binary.

Signed-off-by: Xe Iaso <xe@tailscale.com>