Refactor how we display dates (#56)
Since Next.js server-side renders the page, the dates are parsed in UTC
but then the page is hydrated on the client-side. This was resulting in
incorrect dates. In the cases where we are displaying these dates, we
don't want them to change, so we are now displaying the UTC date.

Resolves #45

Signed-off-by: Tyler Smalley <tyler@tailscale.com>