Align compiler changes with OSS
Tracks tailscale/tailscale@02a2dcfc86eb354e901b5da5df11209eb94845e7

This uses a statically compiled Go toolchain instead of having to fight the
ABI good fight. This also means that I don't have to use a Debian
container for the Go builds and can instead use an alpine:3.17
container. Realistically we could probably get away with alpine:edge to
avoid the time-bomb when Alpine 3.17 eventually goes EOL, but we can
cross that bridge when we come to it.

This also changes the flake.nix to use nixpkgs go to build staticcheck
and whatnot. This shouldn't be a problem in practice.

Signed-off-by: Xe Iaso <xe@tailscale.com>