Extend page background color to overflow scroll (#152)
When applying a background color to a page, it's better to apply it to
the html element rather than the body element, because Firefox and
Chrome use the html background color for the overflow scroll color. It's
a subtle detail, but it makes the page feel more put-together.

This commit also removes classes that were unnecessarily applied twice
from the CSS file and via classes.