Move file to consistent kebab casing (#154)
This commit updates the files in the repository to be consistently
kebab-cased.

I noticed there was a mix of PascalCase files `SEO.tsx`, `Tag.tsx` and
kebab-case `mdx-components.tsx`, `generate-rss.ts`. I suspect this is
the result of multiple developers working on the repo with no clear
standard set. Amongst these two options, kebab case is preferable
because (1) it's what we use for tailscale-www and tailscale-corp and
(2) macOS [doesn't have a case sensitive
file-system](https://twitter.com/kentcdodds/status/1249870276688371713),
which can cause issues when renaming files in Git.

This commit updates all existing files and uses `eslint-plugin-unicorn`
to enforce file name casing in future commits.