scripts: add ElasticSearch janitor (#275)
Our indexing process creates a lot of indices. This adds up and can make
ElasticSearch fall over. This tool cleans up old indices by "closing"
them. It closes every index that is not pointing to one of the aliases
we care about.

This runs every night at 00:42 UTC. The significance of this number is
an exercise for the reader.

Signed-off-by: Xe Iaso <xe@tailscale.com>