components/layout-wrapper: fix header width issues
While I was investigating #289 and #290, I found that there was an
overall issue with header width. After adjusting all of the numbers
to make the header look good on all our supported screen sizes, I
think this should fix the root cause: invisible in Tailwind is not
the same as hidden.

This should fix display on mobile phones as well as make the overall
UX nicer on all browser widths.

Fixes #289
Fixes #290